requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
google-cloud-bigquery==3.20.0
gunicorn==22.0.0